+++ backend/pokerogue/src/main/java/com/pokerogue/helper/battle/BattleMoveRepository.java
@@ -1,5 +1,6 @@
 package com.pokerogue.helper.battle;
 
+import com.pokerogue.helper.battle.data.BattleMove;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
