+++ backend/pokerogue/src/main/java/com/pokerogue/helper/battle/TypeMatchingRepository.java
@@ -1,5 +1,6 @@
 package com.pokerogue.helper.battle;
 
+import com.pokerogue.helper.battle.data.TypeMatching;
 import com.pokerogue.helper.type.data.Type;
 import java.util.HashMap;
 import java.util.List;
