+++ backend/pokerogue/src/main/java/com/pokerogue/helper/battle/controller/BattleController.java
@@ -1,5 +1,9 @@
-package com.pokerogue.helper.battle;
+package com.pokerogue.helper.battle.controller;
 
+import com.pokerogue.helper.battle.dto.BattleResultResponse;
+import com.pokerogue.helper.battle.service.BattleService;
+import com.pokerogue.helper.battle.dto.MoveResponse;
+import com.pokerogue.helper.battle.dto.WeatherResponse;
 import com.pokerogue.helper.util.dto.ApiResponse;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
