+++ backend/pokerogue/src/main/java/com/pokerogue/helper/battle/data/BattleMove.java
@@ -1,5 +1,6 @@
-package com.pokerogue.helper.battle;
+package com.pokerogue.helper.battle.data;
 
+import com.pokerogue.helper.battle.data.MoveCategory;
 import com.pokerogue.helper.type.data.Type;
 
 public record BattleMove(
