+++ backend/pokerogue/src/main/java/com/pokerogue/helper/battle/dto/BattleResultResponse.java
@@ -1,4 +1,4 @@
-package com.pokerogue.helper.battle;
+package com.pokerogue.helper.battle.dto;
 
 public record BattleResultResponse(
         int power,
