+++ backend/pokerogue/src/main/java/com/pokerogue/helper/battle/service/BattleService.java
@@ -1,5 +1,13 @@
-package com.pokerogue.helper.battle;
+package com.pokerogue.helper.battle.service;
 
+import com.pokerogue.helper.battle.data.BattleMove;
+import com.pokerogue.helper.battle.BattleMoveRepository;
+import com.pokerogue.helper.battle.dto.BattleResultResponse;
+import com.pokerogue.helper.battle.dto.MoveResponse;
+import com.pokerogue.helper.battle.TypeMatchingRepository;
+import com.pokerogue.helper.battle.data.Weather;
+import com.pokerogue.helper.battle.dto.WeatherResponse;
+import com.pokerogue.helper.battle.data.TypeMatching;
 import com.pokerogue.helper.global.exception.ErrorMessage;
 import com.pokerogue.helper.global.exception.GlobalCustomException;
 import com.pokerogue.helper.pokemon.data.InMemoryPokemon;
